feat: update metadata for various pages and add sitemap and robots functionality
