+++ src/app/(site)/about-us/page.tsx
@@ -1,5 +1,11 @@
 import { AboutUs } from "./_molecules";
 
+export const metadata = {
+  title: "About ",
+  description:
+    "Learn about I-TAPP's mission to connect Nigerian students with SIWES/industrial training placements, our vision for empowering the next generation of IT professionals, and meet the dedicated team behind our platform.",
+};
+
 export default function AboutUsPage() {
   return <AboutUs />;
 }
