+++ src/app/(site)/contact-us/page.tsx
@@ -1,4 +1,11 @@
+import { Metadata } from "next";
 import { ContactUs } from "./_molecules";
+
+export const metadata: Metadata = {
+  title: "Contact",
+  description:
+    "Get in touch with I-TAPP (ITAPP) for support, partnerships, or inquiries about SIWES and industrial training placements in Nigeria.",
+};
 
 export default function ContactUsPage() {
   return <ContactUs />;
