+++ src/app/robots.ts
@@ -1,0 +1,13 @@
+import type { Metadata, MetadataRoute } from "next";
+
+export default function robots(): MetadataRoute.Robots {
+  const siteUrl = "https://www.i-tapp.com";
+
+  return {
+    rules: {
+      userAgent: "*",
+      allow: "/",
+    },
+    sitemap: `${siteUrl}/sitemap.xml`,
+  };
+}
